feat: move admin WebSocket wire types into shared/wire.ts Wire namespace
- Create shared/wire.ts as canonical home for all wire types (both admin
  WS types and foreman↔worker protocol types from src/wire.ts)
- Rename TaskSnapshot → Wire.Task, WorkerSnapshot → Wire.Worker per
  concept-named convention; keep LogEntry, BlockerInfo, AdminMessage
- Make src/wire.ts a re-export shim for backward compatibility
- Update all consumers: foreman models, admin-ws.ts, frontend types.ts,
  frontend pages, and tests/admin-ws.test.ts
- Update CLAUDE.md and docs/type-system.md to reflect new location

Closes #630
